koviz core dump on single big trk datafile
This is due to using an int rather than a qint64 for
the underlying TrickModelIterator counter.

I went to an int because I was trying to get rid of compiler
warnings due to comparing an int and a qint64.

+++ libkoviz/trickmodel.h
@@ -306,7 +306,7 @@ class TrickModelIterator : public ModelIterator
 
   private:
 
-    int i;
+    qint64 i;
     const TrickModel* _model;
     int _row_count;
     int _row_size;
